pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "excel-translator"
version = "0.1.0"
description = "ゲームテキスト翻訳システム - Excel to LLM Translation Pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Developer"}
]
keywords = ["translation", "excel", "llm", "game-localization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "google-generativeai>=0.4.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.98.0",
]

[project.scripts]
excel-translator = "excel_translator.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.hypothesis]
max_examples = 100
